+++ sc-kpm/sc-agents-common/utils/AgentUtils.hpp
@@ -26,31 +26,10 @@ public:
       sc_uint32 const & waitTime = DEFAULT_WAIT_TIME,
       ScAddr onEventClassAddr = ScAddr::Empty);
 
-  SC_DEPRECATED(
-      0.9.0,
-      "Use ScAddr applyActionAndGetResultIfExists"
-      "ScMemoryContext * ms_context, const ScAddr & actionClass, const ScAddrVector & params, "
-      "const sc_uint32 & waitTime) instead of. It will be removed in sc-machine 0.10.0.")
-  static ScAddr getActionResultIfExists(
-      ScMemoryContext * ms_context,
-      ScAddr const & actionClass,
-      ScAddrVector const & params,
-      sc_uint32 const & waitTime = DEFAULT_WAIT_TIME);
-
   static ScAddr applyActionAndGetResultIfExists(
       ScMemoryContext * ms_context,
       ScAddr const & actionClass,
       ScAddrVector const & params,
-      sc_uint32 const & waitTime = DEFAULT_WAIT_TIME);
-
-  SC_DEPRECATED(
-      0.9.0,
-      "Use ScAddr applyActionAndGetResultIfExists"
-      "ScMemoryContext * ms_context, const ScAddr & actionNode, const sc_uint32 & waitTime) instead of. "
-      "It will be removed in sc-machine 0.10.0.")
-  static ScAddr getActionResultIfExists(
-      ScMemoryContext * ms_context,
-      ScAddr const & actionNode,
       sc_uint32 const & waitTime = DEFAULT_WAIT_TIME);
 
   static ScAddr applyActionAndGetResultIfExists(