+++ sc-memory/sc-memory/sc_template.cpp
@@ -14,14 +14,10 @@ extern "C"
 #include "sc_template.hpp"
 #include "sc_memory.hpp"
 
-SC_DEPRECATED(0.8.0, "Don't use alias for fixed sc-address")
-
 ScTemplateItem operator>>(ScAddr const & value, char const * replName)
 {
   return {value, replName};
 }
-
-SC_DEPRECATED(0.8.0, "Don't use alias for fixed sc-address")
 
 ScTemplateItem operator>>(ScAddr const & value, std::string const & replName)
 {